requirements.txt
python-telegram-bot
jinja2
weasyprint
jdatetime
